Add tests. Refactor ical view
